Bump json-schema version to 2.2.3
This is to pick up our fix to the gem for symbol keys in hashes in
arrays:

    https://github.com/hoxworth/json-schema/pull/105

This means that we can write our test data more idiomatically as ruby
hashes with symbol keys. While working on the schema we had found a
workaround to this which involved calling .to_json on the data before
validating it, but the fix makes this unnecessary.