Add examples of JSON to be sent to the API
These are based on what we think we need to send to the content-store.
Some fields, such as a section's manual slug, will be used to generate
base_path and routes and will not be included in the data sent to the
content-store.

The content of these examples is derived from the scraped JSON in
manuals-frontend. We have simplified it to make it shorter and easier
to read, and the relationships between sections and the manual are not
consistent (eg the sections in EIM25525's breadcrumbs are not included
here).