Remove titles from breadcrumbs
We added this speculatively because we thought it might allow a better user
experience when looking at a long chain of breadcrumbs - it could be included
as the link's title.

Feedback from HMRC/their development team at iO1 is that updating a section
title near the top of the hierarchy could trigger hundreds of requests to
update every child section in the tree. The trade-off doesn't feel worth it.