Remove manual slug from Section schema
This information is already included in the URL they PUTed to, so it seems
unnecessary to require it. We still need to send it to Publishing API/Content
Store so that a section document knows which manual it is a part of, without
having to know about the structure of its base_path.