Be more specific with section slugs versus IDs
It is coincidental that section IDs are used for section IDs. This may not
always be the case, and we shouldn't build dependence upon this.

Additionally, when they are published on GOV.UK, the section slugs will be
lower case, not uppercase, so there is already a distinction. We may decide to
mirror them being lowercase in our API.