Slugs in request URLs must be lowercase
The GOV.UK slugs/URLs will be lowercase, so let's enforce it to avoid a
situation where they send it uppercase but we neglect to downcase it.

We will still need to remember to downcase Section IDs in breadcrumbs and
child_section_groups before we make them into slugs as part of base_paths.