Push manuals into the search index
In the medium term, the search index will automatically be populated when they
are published to Content Store via the Publishing API. For now, though, we need
to push them directly. The same transformation we perform will need to be
performed in the code which populates the search index, so this code may be
reused.