Better section titles in search results
At the moment, sections appearing in search results don't identify themselves
very well because they are just the section title itself, which, based on the
existing HMRC manuals on www.hmrc.gov.uk, would be things like "Definition of a
Chapter 2 lease". With this change, you can see that it is an HMRC Manual, and
if you searched by section ID, you see that in the search result to help you
identify it.