Include the format in search documents
We should include the format of the document so that it can be used in the
search index for various things, such as weighting the format up and down.

More specifically/imminently, we want to display the format in search results
to give users context on what the search result is when HMRC Manuals and
sections appear in search results.