Return the status code from the Publishing API to the client
In production, Rummager queues search index updates so returns 202
to POST requests to insert or overwrite a document:

https://github.com/alphagov/rummager/blob/cc0e9c95db9098b0dbc3dd12124d1eef5a153ae7/app.rb#L460-L466

We have been returning the status code from Rummager to the client,
which is inconsistent with the documentation in the README. It also
seems inaccurate: the content has been saved to the content-store
already and so can be displayed on GOV.UK, so 200/201 is more appropriate.

This commit switches to returning the Publishing API response status
code to the client. It also bumps the version of gds-api-adapters to
pick up the addition of a test helper which stubs a POST request to
Rummager as if queueing is enabled.