Make section titles in search results less repetitive
This was included to provide some context/additional information about the
search result is as the titles of sections are often obscure and sections have
no description.

Now, though, the format name is displayed underneath HMRC manual search
results, so that rationale is no longer there. Instead, having the format name
in the search result title makes the list noisier, especially if there are
multiple HMRC manual results.