Include leading slashes in several fields sent to Rummager
- The link and ID fields in manuals sent to Rummager should be the
  full path, but currently it's the slug without the leading slash.
- Do the same for manual sections link and ID fields.

The erroneous behaviour that we're fixing was originally thought to have
been copied from Specialist Publisher and only be working now by
coincidence. Now it turns out that when we make manuals appear on topic
pages the links to the manuals will be broken if we don't fix this.
(That's because so far the pages which fetch their content from Rummager
and include links to manuals have only a single path segment (eg
/search), so treating the manual link as relative to the current URL
does the right thing from those pages. Topic pages have longer paths (eg
/topic/business-tax/vat) so a relative manual link from those pages will
be incorrect (eg
/topic/business-tax/hmrc-internal-manuals/vat-government-and-public-bodies).)

_Words in the latter set of parentheses written by Jenny more eloquently
than I could!_

Trello: https://trello.com/c/xhObMTuI/72-all-manuals-are-sent-to-rummager-without-the-leading-slash-on-the-path-in-the-link-field.