Load manuals-to-topics mapping in initializer
Creates a hash based on the 'manuals_to_topics' CSV, mapping manual slugs to
the `content_id`s of their topics, like

  "manual_slug" => ["a", "b", "c"]

This is a step towards being able to tell Publishing API and Rummager about the
topics a Manual should be linked to.

We're dealing with the topics' `content_id`s because they're intended to be
more stable than slugs. (You may notice that the topics' slugs are also
included in the CSV, but this is merely intended to be used as an annotation to
make the file _slightly_ more human-readable.) Where slugs are needed, the
intention is to use their `content_id` to fetch them on-the-fly from Content
Store.