Load manuals-to-topics mapping in initializer
Creates a hash based on the 'manuals_to_topics' CSV, mapping manual slugs to
the `content_id`s of their topics, like

  "manual_slug" => ["a", "b", "c"]

This is a step towards being able to tell Publishing API and Rummager about the
topics a Manual should be linked to.

The naming of the objects involved really is not great. Suggestions welcome!