Link topics in PublishingAPIManual hash
"There is currently no way for HMRC to tag manuals to topics (ie, the hierarchy
which now lives under https://www.gov.uk/topic/). This means that manuals
cannot appear on these pages, which will make them hard for users to find."

As the name suggests, the PublishingAPIManual model exports a representation of
a Manual to be consumed by the Publishing API. Including a list of topics in
that representation is a good step towards tagging Manuals to topics.

Putting the topics list in the `links` field was suggested as a New World
alternative to using the the `details` hash's `tags` field, as is currently
expected by services like the Email Alert Service/API. Therefore, changes will
be required in other apps before this means anything.

I don't fully understand the migration plan from 0 manuals with topics to 100%
of manuals with topics, so I figured optionally including the topics list for
now would be easiest.