Send latest change note as string
With the work completed in 729effb we were incorrectly sending a hash to
Rummager for the latest change note. This commit updates it to send a string
structured like so "Description of change in Title of section changed."

The format was chosen based on feedback from the Content team ("An em dash
isn't appropriate here, and since colons are already used frequently in section
titles they aren't a good option either in this context").

This format won't be right for longer full-sentence change notes but most so far
(https://www.gov.uk/hmrc-internal-manuals/vat-government-and-public-bodies/updates)
have been short phrases and we can change the format relatively easily when
we've seen a few more manuals. (When changing the format, manuals will need
republishing to resend the change note to Rummager.)