Fix leading slash script for change notes
Now that we're sending the latest change note to Rummager, the script to fix
the leading slashes needs to include the details hash in the data passed to
the RummagerManual instance. Since this is a one-time script and will be run
straight after deploying 729effb1add96cfea09255fb558a355b19125daa we won't
lose any data by using an empty array for the change_notes field here.