Put `maximal_manual`'s slug, etc in code
There are some 'test data' modules defining and using the idea of a 'maximal
manual' and a 'maximal section' - I assume they're just a manual and section
_with everything_? These maximal objects' slug, base_path and URL were not defined in any
of these modules, which means

  1. they're hard-coded and repeated in tests with little explanation of what
     they are, and
  2. we can't reuse them very pleasantly in a different context (say, in the
     test data modules themselves, even)

I've chosen not to deduplicate the repeated sections of these strings (i.e. an
object's url contains its base_path which contains its slug) so they're more
easily readable and greppable. I have grouped them all close together at the
top of their file, though, to make this visible.

I'm not sure whether these should live in the TestDataHelpers or the
PublishingApiDataHelpers, so I went with the former since it sounds the most
generic and might encourage these to be reused in more situations.