Add model to represent removing a manual
Some manuals have been published to test the system.  Now we know the system works we'd like to remove them.  This commit is step 1 in allowing us to automate that removal.

We model this as a new object that represents a request to remove a manual.  It may have made sense to add remove functionality to the existing PublishingAPIManual object, but it didn't really feel like the functionality fitted in with that model due to the validation concerns and choice of #to_h as the interface to convert into the json payload for publishing-api.