Add calls to Rummager to delete manuals
When we remove a HMRC Manual we also remove it from Rummager so it does not appear in Search.
In the model specs, we have put a placeholder to assert that the Rummager delete request has been made, as the current rummager helper method does not perform an assertion. There is an open PR (https://github.com/alphagov/gds-api-adapters/pull/362) for the `assert_rummager_deleted_item` which addresses this.