Allow HMRC to send content_id
This commit makes it possible for HMRC to send a `content_id` for
manuals and sections. We'd like manuals to have content_ids to enable
tagging in the new tagging infrastructure.

The tests require a little explanation. By adding a `content_id` to
`TestDataHelpers#maximal_manual` we simulate all calls to `PUT
/hmrc-manuals-api/:slug` in `requests/manual_spec.rb` includea
content_id.

By adding the same `content_id` to
`PublishingApiDataHelpers#maximal_manual_for_publishing_api` we then
verify that the content_id is forwarded to the content-store.