Removed Manuals are invalid if they don't exist
This is to avoid us publishing "gone" items over urls that didn't already exist.  After some debate we decided that "existing" here means that we can ask content store for the item.  If we get an item that matches the supplied format then we consider it to exist.  If it's not an item, or is already a "gone" item then we consider it to not exist and don't let the removed manual make changes.

Note that we only run this validation if there are no errors on the :slug as otherwise we run the risk of trying to fetch broken URIs from the API.  This relies on validation for the slug happening before this validation.