Generate a section uuid from base_path...
Generates a RFC4122 compliant v5 uuid from the base_path of the manual,
and use this value where the content_id is missing.
This allows us to predictably assign content_ids which pass Publishing-API
uuid validations and are consistent for the same base_path.

TBC: Should the same content_id exist for paired manuals and sections?