Generate uuid from base_path...
Generates a RFC4122 compliant v5 uuid from the base_path of the manual
or section.
We only use this uuid where the content_id is missing.
This allows us to predictably assign content_ids which pass Publishing-API
uuid validations and are consistent for the same base_path and have a
method which HMRC can use to predictably generate uuids in future.