Use v2 of publishing-api
All apps should now use v2 of the publishing api to write content to the content-store. This
change converts hmrc-manuals-api to use v2 of the api. The differences are:

- put content_item is replaced by three methods - put_content, publish, and put_links
- the content_id is used as the key, rather than the base path
- the base path is included in the payload, links are not included in the payload of the put_content

The three requests to publishing-api have been refactored out into a PublishingAPINotifier class
which halts processing if any of the responses are non-200, and returns the response to the client.

Trello: https://trello.com/c/II6tydnO/453-use-publishing-api-v2-for-hmrc-manuals-api