Withdraw manual after its sections
We need to withdraw some large manuals with over 2000 sections. If the task
exits for some reason before it has finished, the manual will have been
withdrawn already and so a second attempt to run the task will error because
the manual is no longer in the content store. Switch to withdrawing the manual
last to avoid this.