Update README following styleguide
We move the current overly detailed README contents to a file in the 'docs' folder and create a new README that is more consistent with our [style guide](https://github.com/alphagov/styleguides/blob/master/use-of-READMEs.md). We link to appropriate header tags from the new README to the more detailed one.

We add the missing LICENCE file.