Use shoulda-matchers
The testing code for validations was quite verbose, and could be shortened with using the 'shoulda-matchers' gem. The matchers work by setting a given instance variable and running `.valid?`, so we need to change certain variables from `attr_reader` to `attr_accessor`.

For `PublishingApiManual` and `PublishingApiSection`, there were no format validation tests so we added them. As part of the other 'to_h' validation, the class method `base_path` is called, which errors when we set the slug/s to `nil`. Therefore we change the `base_path` method to force a slug to be a string in order to make the tests work.