Add script to check for orphaned sections for a manual or section
We receive requests from HMRC for the removal of manuals or sections.
Before we do this it is necessary to check whether any sections that
belong to the thing being removed has been given another parent setion
or manual. This avoids any sections becoming orphans.

HMRC should be able to do this themselves before raising a request with
us. This script checks for any sections that have not been re assigned
to another parent and lists those that require changing.

[Trello
card](https://trello.com/c/cIESLI1q/463-write-a-script-to-automate-whether-an-hmrc-manuals-section-can-be-deleted-3)