Allow sections to be redirected to parent manual
Added two rake tasks: `redirect_hmrc_section_to_parent_manual` and
`redirect_all_hrmc_sections_to_parent_manual`.  At present we don't have any need
to redirect sections to other manuals or manuals to other manuals.

The `redirect_all_hrmc_sections_to_parent_manual' task outputs the source and
destination slugs which could be handy if something went awry.

Added a new model `PublishingAPIRedirectedSectionToParentManual` in the same
vein as `PublishingAPIRedirectedSection`.  There seems to be enough difference
in the validations (for example) to keep it separate.

Added a `SectionRetriever` class to encapsulate the construction and execution
of Rummager queries, and allow construction of result items as required so we
can construct both the `Redirected` and `Removed` section types from it.
I tried shoehorning it into `PublishingAPIManual`, but it didn't seem to fit
very nicely.  We don't need it for anything that PublishingAPIManual does at
the moment, so it seemed wrong to add complexity there.