Refactor PublishingApiRedirectedSectionToParentManual
Now that PublishingApiRedirectedSection allows redirecting to a manual
as well as to a section, the PublishingApiRedirectedSectionToParentlManual
is just a specialization that takes two arguments instead of three (the
third argument is just the first one repeated).

I thought about removing this entirely, but the `.from_rummager_result`
method meant there are slightly different usecases and so I left it in.