Refactor authentication spec for gds-sso 13.2.0
We now determine an api call by the presence of a bearer token,
if this isn't present the default warden strategy is used
(in test mode this takes the first user from the db).
With the mock user disabled the default strategy redirects
the user to /auth/gds.
Despite this I'm leaving this test in place to highlight
the problem of a broken redirect.