Change update_type to nil for publish requests
We are deprecating sending the `update_type` in publish requests made to the
Publishing API, and now expect clients to send `update_type` in PUT requests
instead.  In this app we are currently passing 3 params to publish, so we must
continue to send all 3 arguments until `update_type` is fully deprecated, hence
why we change it to `nil` in the meantime.