Use unicorn to allow concurrent connections
The app is currently using `rails s` as a server. This does not allow
concurrent connections.

In the event that a request is slow, it is possible the next request
will not be possible and nginx will timeout.

By switching to use `govuk_unicorn`, we can accept two concurrent
connections which may help alleviate the issue.