Remove sprockets-rails gem
This application does not need this gem as it doesn't have assets. The
presence of it actually causes problems with the govspeak gem (which
depends on govuk_publishing_components which depends on Sprockets) as
with a Sprockets 4 dependency, the presence of this gem requires an
asset manifest file - which, of course, is unnecessary if you don't have
assets.