Set phase to 'beta' for HMRC manuals
Currently HMRC manuals / sections have a 'beta' banner when they're
rendered by Manuals Frontend. This is done via some custom logic
depending on whether the manual is from HMRC. However, the correct way
to indicate a content item's phase is to set the 'phase' attribute on
the content item instead.

We need to do this as we migrate from Manuals Frontend to Government
Frontend.

Trello:
https://trello.com/c/YmmLFXu5/1127-manuals-migrate-manuals-frontend-code-to-government-frontend