Allow previously redirected manuals to be removed
If a manual or a section has been previously redirected, it has a schema
of `redirect`. This means we cannot subsequently remove the manual or
the redirect, as the new content item won't be valid.

Therefore allowing a schema of `redirect` for us to be able to remove a
manual or a section that has previously been redirected.