Return 422 when request is unprocessable by Publishing API
If the user provides data that is unprocessable by Publishing API (e.g.
a base path with the incorrect content ID), we currently return a 500
error and log this to Sentry.

This is not optimal behaviour as there is no application fault here, so
no action can be taken by developers after seeing this logged in Sentry.

Instead we should respond with a 422 and return the actual error to the
user. Then they can correct the request they are making so it becomes
valid.